geocoder

+++ internal/parser/models/shop.go
@@ -1,8 +1,8 @@
 package models
 
 type Coordinates struct {
-	Latitude  float32 `json:"latitude"`
-	Longitude float32 `json:"longitude"`
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
 }
 
 type Shop struct {
